design(web): operator console lives inside the app shell
Move the console under the collapsible sidebar shell as a routed page
(ConsolePage) with a redesigned header (ConsoleHeader: crumb, status pill, spend,
pause/resume/stop, New run, report, and the add/remove-panel layout menu) in
place of the old topbar buttons. Extract NewRunDialog; drop the standalone
SessionShell/SessionTopbar. Console renders in the floating card and the sidebar
collapses like every other page.
