+++ apps/web/src/app/ConsolePage.tsx
@@ -3,10 +3,9 @@ import { useParams } from 'react-router-dom';
 import { useUI } from '@/store/ui';
 import { useSession } from '@/features/hooks';
 import { Spinner } from '@/components/ui/primitives';
-import { SessionTopbar } from './SessionTopbar';
 import { Workspace } from './Workspace';
 
-export function SessionShell() {
+export function ConsolePage() {
   const { id } = useParams();
   const setActiveSession = useUI((s) => s.setActiveSession);
   const setActiveRun = useUI((s) => s.setActiveRun);
@@ -20,18 +19,12 @@ export function SessionShell() {
     if (session) setActiveRun(session.activeRunId ?? null);
   }, [session, setActiveRun]);
 
-  return (
-    <div className="flex h-screen w-screen flex-col overflow-hidden bg-bg">
-      <SessionTopbar session={session ?? null} />
-      <div className="min-h-0 flex-1">
-        {isLoading && !session ? (
-          <div className="grid h-full place-items-center">
-            <Spinner />
-          </div>
-        ) : (
-          <Workspace />
-        )}
+  if (isLoading && !session) {
+    return (
+      <div className="grid h-full place-items-center">
+        <Spinner />
       </div>
-    </div>
-  );
+    );
+  }
+  return <Workspace />;
 }
